Documentation report generator
This patch introduces a new generator that produces a report of the
status of the documentation of the model. The report includes the names
of types and services and a score calculate dividing the number of
documented items by the total number of items.

The report is generated in CSV format, so that it can be easily loaded
into an spreadsheet and used to generate more user friendly reports.

Change-Id: I61709e4ce9a7fe80ca3e921640dcee46817a7d1f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 9679bfd4f6cf59cb4fa663129442270e4f3a0b82)