Name element tag by the enitity, which is written
When writing XML, we always written XML element same as member,
which we wanted to write. This patch changes it, so we name the
singular element tag same as entity we write, not the member.

For example if we want to create bonds, we will send following xml:

  <action>
    <modified_bonds>
      <host_nic>
        <bonding>
          <options>...</options>
          <slaves>
            <host_nic>
              <name>eth0</name>
            </host_nic>
            <host_nic>
              <name>eth1</name>
            </host_nic>
          </slaves>
        </bonding>
        <name>bond0</name>
      </host_nic>
    </modified_bonds>
  </action>

Previously we sent <slave> instead of <host_nic> in <slaves> element.

Change-Id: I875ec6c1b77ece7f26315833c1032c7dd62b535e
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit 5c8ba0d038f957b637174b261263cfb0634de94c)