Support custom AsciiDoc attributes
Currently the set of AsciiDoc attributes used when generating and
rendering AsciiDoc attributes is hardcoded, and can't be changed by
users of the tools.  This complicates, for example, generation of
customized versions of the documentation. To make that easier, this
patch adds to the tool a new '--adoc-attribute' option that can be used
to specify custom AsciiDoc attributes. With this the developer of the
model can, for example, use a placeholder like '{myvariable}' in the
text, and then, when using the tool replace that with a custom value,
simply adding the following option:

  --adoc-attribute='myvariable:myvalue'

These custom attributes will be added to the header of the AsciiDoc text
generated from the model, and will also be used when generating the
fragments of HTML that are included in the XML and JSON representations
of the model.

Change-Id: Ia52b1404350de99d90823bb06f880a40983ae000
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 6bcd3f9a1369e135e6998f7119b8e6c522f637a9)