Make section id separator configurable
Currently the character used to separate the segments of section
identifiers is the dash. For example, the section identifier for "Vm"
type would be like this:

  types/vm

This was changed recently to simplify the use of the generated .adoc
files in the documentation build system used by Red Hat. But this
separator is also used in the cross references. For example, when a pice
of documentation wants to refer to the "Vm" type it can use a cross
reference like this:

  ... see the <<types/vm, Vm>> type for ...

Those cross references haven't been changed, so they are now incorrect.
To address that issue without having to modify all the cross references
manually this patch modifies the metamodel tool so that this character
is configurable. The tool will also automatically replace the slash with
the selected character in all the cross references.

The value can be changed using the "adoc.separator" build property. For
example, to use the dash as the separator:

  $ mvn validate -Pdocument -Dadoc.separator=-

Change-Id: I38bbbba788e4dff31586efdbcd2b08b34640f79c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 74100caf0d2b654807bb023eea9ab3e11f6212e6)