Avoid duplicate section identifiers
Currently the generator of the AsciiDoc documentation generates the
section identifiers for attributes of types concatenating the name of
the type and the name of the attribute. For example, the identifier of
the section of the "Vm.memory" attribute should be:

  types/vm/attributes/memory

But the "Vm" type is an extension of the "VmBase" type, and the "memory"
attribute is declared in there, in "VmBase". So the actual generated
section identifier is this:

  types/vm_base/attributes/memory

The "memory" attribute is then used in multiple places, in particular it
is used in the "Vm" and "Template" types. The net result is that the
section identifier is duplicated. This causes problems in some tools, in
particular in the tools used by Red Hat to generate their documentation.

The same happens with methods of services.

To avoid that issue this patch changes the tool so that it generates
correctly the section identifiers, using the name of the derived type,
not the name of the base type.

Change-Id: I94e14f537198f0a6dbd0565d9130e2da06134fb9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit daf45da39fe36f6307fcdc16d4c65803df15795a)