model backward compatibility - interface default methods
When one wishes to expose a new api, a patch needs to be pushed
to the model project with the api definition. On the next model release
it will include the new api.
When building the ovirt-engine, interfaces are generated automatically
from the current model.
The engine implementation patches can't be merged before the model
version is released, as we the generated interface won't include the method
(unless we dismiss @Override which is undesired) which will break the build.
When the new model version is released, the engine master won't compile
till all the new interface methods implementation will be merged.

This patch changes the metamodel toi leverage java 8 default methods -
the interfaces will be generated with default methods mainting backward
compatibility with previous implementations of the interfaces.

Change-Id: I8f7f88069242abc6c6735b8db84ca7562b52f901
(cherry picked from commit ca81245883c8081f2edd461ad46a75bfee3bbd3e)