Use non capturing groups in @Path annotations
Currently the @Path annotations generated for the resources used to
check the status of asynchronous actions is similar to this:

 @Path("{action: (import|register)}/{oid}")
 ActionResource getActionResource(
   @PathParam("action") String action,
   @PathParam("oid") String oid
 )

But in version 3 of Resteasy the use of parenthesis and resource
locators doesn't work correctly. See this issue for details:

  Regex with parenthesis doesn't work with resource locator
  https://issues.jboss.org/browse/RESTEASY-1569

A simple workaround for that is to use a '(?:' instead of '('. That
means that the regular expression will use a non capturing group instead
of a capturing group. For our use case this is perfectly valid, as we
don't need a capturing group. This patch applies that workaround.

Change-Id: I5dc1f6e683a973bfbd08bee6450ca7ec1f0db149
Related-To: https://bugzilla.redhat.com/1405774
Related-To: https://issues.jboss.org/browse/RESTEASY-1569
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 9c9982e430dcda819fd9d6f9bcc631bae15ec896)
(cherry picked from commit 42b6d59a38fc6ba56e047bc04f7ad4d6c221c5de)