Fix reading XML attributes of enum types
Currently the generated XML support code doesn't read correctly the
values of model attributes that are represented as XML attributes. For
example, if a type has an 'id' attribute of a enum type, it will just be
ignored. This patch fixes the XML support code generator so that this
works correctly. The patch also adds a couple of tests to verify that
this kind of attributes are read and written correctly.

Change-Id: I1cb2599802b5b6a5e2c877896089865e7f37e7a8
Related-To: https://bugzilla.redhat.com/1408839
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 6231ef85a4118ffff7fab0b3bb6103e1399538e9)