Check AsciiDoc cross references
This patch modifies the AsciiDoc generator so that it checks that there
are no cross references without a corresponding section identifier. This
is needed because the some publication tools consider that an error, and
we need to catch it earlier. With this change the build will fail, and
we will detect the error before merging the offending patch.

Change-Id: I170eba43afd9eef91d60483f5289ed13a071ccd8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit efe165176f78025df70942129dcd1a322a960975)