Fix XML tag for lists of enums
Currently the code generated for XML writers doesn't create the correct
XML for lists of enum values. For example, for the `Boot.devices`
attribute it generates the following:

  <devices>
    <boot_device>...</boot_device>
    <boot_device>...</boot_device>
    ...
  </devices>

But it should generate what the server expects, which is this:

  <devices>
    <device>...</device>
    <device>...</device>
    ...
  </devices>

The rule is that the name of the tag for each element is the singular of
the name of the attribute.

Change-Id: I7631e3a2772e6f067fbdac5182e291472f4c47a4
Bug-Url: https://bugzilla.redhat.com/1512850
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 7600cc0d9244758349674cf548d11560242a2476)