Generated Exception for Sub-Resource loacation
When adding a new CRUD operation or an Action to an API
service in the model, at the onset it is naturally missing
the implmenetaion in the Engine.

To prevent the Engine from breaking on compilation, methods
are generated with 'default' keyword and a default
implementation which throws and UnsupportedOperationException.

For some reason this functionality is missing in sub-resource
locator methods; i.e, methods in a resource which return
one of its sub-resources.

This results in Engine breaking on compilation when a new
sub-resource, "branch in the API tree" is added to the API.

This patch adds this missing functionality

Change-Id: I3dffdbf73b244376e5a67d111110fe231064b571
Signed-off-by: Ori_Liel <oliel@redhat.com>