Handle Enum URL Parameter
As of ovirt-engine-api-model 4.4.18 the API contains the
first instance of a URL parameter which is an enum
(AutoPinningPolicy).

The metamodel is unfamiliar to such a possibilty and
the generated code fails compilation.

This patch adds a default rendering method renderOther()
which deafults to .toString for URL parameters which are
not of the standard types: String, Integer, Boolean,
Decimal, Date

Change-Id: I819e4c45e598c68e758e86d13e5f8530cb9d1f24
Signed-off-by: Ori_Liel <oliel@redhat.com>